Add files via upload
In this update I added the music for the aliens which interrupts the broadcasts and the final car ride. I also combined the sections of the music into one final song at the end. I included updates to the sources document, new screenshots of my work on the songs, and the background for my work. Note that the soundscape is best listened to with headphones.